# docs/swagger.yaml
definitions:
  apis.ErrResponse:
    description: Структура HTTP ответа об ошибке
    properties:
      code:
        description: application-specific error code
        type: integer
      error:
        description: application-level error message, for debugging
        type: string
      status:
        description: user-level status message
        type: string
    type: object
  apis.Response:
    description: 'Структура HTTP ответа: если отправка успешна, дополнительно возвращается
      id вставленной записи.'
    properties:
      id:
        example: "123"
        type: string
      message:
        example: OK
        type: string
    type: object
  models.Images:
    properties:
      title:
        example: Спуск. Фото №99
        type: string
      url:
        example: https://avatars.mds.yandex.net/i?id=a467876d3e1b1f0a84050103a206cf81-5858922-images-thumbs&n=13
        type: string
    type: object
  models.Pereval:
    properties:
      add_time:
        example: "2021-09-22 13:18:13"
        type: string
      beautyTitle:
        example: 'пер. '
        type: string
      connect:
        example: ' '
        type: string
      coords:
        properties:
          height:
            example: "1200"
            type: string
          latitude:
            example: "45.3842"
            type: string
          longitude:
            example: "7.1525"
            type: string
        type: object
      images:
        items:
          $ref: '#/definitions/models.Images'
        type: array
      level:
        properties:
          autumn:
            example: 1A
            type: string
          spring:
            example: ' '
            type: string
          summer:
            example: 1A
            type: string
          winter:
            example: ' '
            type: string
        type: object
      other_titles:
        example: "1"
        type: string
      pereval_id:
        example: "125"
        type: string
      title:
        example: Туя-Ашуу
        type: string
      type:
        example: pass
        type: string
      user:
        $ref: '#/definitions/models.User'
    type: object
  models.User:
    properties:
      email:
        example: dd@aa.ru
        type: string
      fam:
        example: Скворцов
        type: string
      id:
        example: "11234"
        type: string
      name:
        example: Иван
        type: string
      otc:
        example: Кожедубович
        type: string
      phone:
        example: "+744434555"
        type: string
    type: object
info:
  contact:
    email: serj_bibox@mail.ru
    name: API Support
  description: API для взаимодействия приложения с сервером БД ФСТР.
  title: ФСТР API
  version: "1.0"
paths:
  /submitData:
    post:
      consumes:
      - application/json
      parameters:
      - description: карточка объекта
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Pereval'
      - description: ID созданной записи
        in: body
        name: output
        required: true
        schema:
          $ref: '#/definitions/apis.Response'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apis.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apis.ErrResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/apis.ErrResponse'
      summary: Создаёт новую запись в pereval_added
      tags:
      - /submitData
  /submitData/:id:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/apis.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apis.ErrResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/apis.ErrResponse'
      summary: Получает запись из pereval_added по ID записи
      tags:
      - /submitData/:id
swagger: "2.0"